+++ client/src/main/java/com/paraguacraft/pvp/ParaguacraftPvP.java
@@ -36,7 +36,7 @@ public class ParaguacraftPvP {
 
     public static final String MODID = "paraguacraftpvp";
     public static final String NAME = "Paraguacraft PvP Client";
-    public static final String VERSION = "2.1.31";
+    public static final String VERSION = "2.1.35";
 
     @Mod.Instance(MODID)
     public static ParaguacraftPvP instance;
@@ -59,6 +59,7 @@ public class ParaguacraftPvP {
         MinecraftForge.EVENT_BUS.register(new QoLManager());
         MinecraftForge.EVENT_BUS.register(new GuiBackgroundHandler());
         MinecraftForge.EVENT_BUS.register(new com.paraguacraft.pvp.modules.VisualsManager());
+        MinecraftForge.EVENT_BUS.register(new com.paraguacraft.pvp.modules.TntCountdownRenderer());
         MinecraftForge.EVENT_BUS.register(new BadgeNetHandler());
         MinecraftForge.EVENT_BUS.register(new DiscordPresenceHandler());
         MinecraftForge.EVENT_BUS.register(new LauncherIpcHandler());